Improve the ability to explain Uxntal words.
This commit unifies the previous support for explaining
Uxntal instructions and numeric constants to include most
other syntax as well. Rather than having to choose the
correct commmand to run for a given word, we reuse the
regular expressions used for syntax highlighting.

Additionally, the file has been restrucuted to be more
compliant with MELPA requirements.